requirement.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
pypdf
python-dotenv
streamlit